refactor(frontend): extract callout archive panel from CNR list page
